requirements.txt
azure-eventhub>=5.11.0
azure-eventhub-checkpointstoreblob-aio>=1.1.4
python-dotenv>=1.0.0
aiohttp>=3.8.0
asyncio>=3.4.3
aiofiles>=23.1.0
